requirements.txt
aioredis==1.3.1
aniso8601==7.0.0
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
autopep8==1.5.7
boto==2.49.0
boto3==1.18.6
botocore==1.21.6
certifi==2021.5.30
cffi==1.14.6
channels==3.0.4
channels-redis==3.3.0
charset-normalizer==2.0.3
colorama==0.4.4
constantly==15.1.0
cryptography==3.4.7
daphne==3.0.2
decorator==4.4.2
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.5
django-allauth==0.45.0
django-cleanup==5.2.0
django-cors-headers==3.7.0
django-debug-toolbar==3.2.1
django-filter==2.4.0
django-graphql-auth==0.3.16
django-graphql-jwt==0.3.0
django-pwa==1.0.10
django-rest-auth==0.9.5
django-storages==1.11.1
django-utils-six==2.0
djangorestframework==3.12.4
graphene==2.1.9
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
gunicorn==20.1.0
hiredis==2.0.0
hyperlink==21.0.0
idna==3.2
imageio==2.9.0
imageio-ffmpeg==0.4.4
incremental==21.3.0
isort==5.9.2
jmespath==0.10.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
moviepy==1.0.3
msgpack==1.0.2
ndg-httpsclient==0.5.1
numpy==1.21.1
oauthlib==3.1.1
Pillow==8.3.1
ply==3.11
proglog==0.1.9
promise==2.3
psycopg2==2.9.1
psycopg2-binary==2.9.1
pusher==3.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
PyJWT==1.7.1
PyNaCl==1.4.0
pyOpenSSL==20.0.1
python-dateutil==2.8.2
python-decouple==3.4
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
Rx==1.6.1
s3transfer==0.5.0
service-identity==21.1.0
singledispatch==3.6.2
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
tqdm==4.61.2
Twisted==21.2.0
# twisted-iocpsupport==1.0.1
txaio==21.2.1
urllib3==1.26.6
whitenoise==5.3.0
zope.interface==5.4.0
django-compressor==2.4.1
